Revert "Refactor Streamlit app modules"
